# API/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Citi+
  description: Documentation for Citi+ REST API.
  termsOfService: https://www.ratp.fr/gestion-de-vos-donnees-personnelles
  contact:
    email: l_laugier@etu-webschoolfactory.fr
  license:
    name: MIT License
    url: http://localhost:80/api/v1/documentation/LICENSE
  version: 1.0.0
...
servers:
- url: http://localhost:80/api/v1/
...
paths:
...
  /espaces-verts:
    get:
      tags:
      - paris
      summary: Get all green areas of Paris
      description: test
      operationId: GreenAreasGET
      parameters:
      - name: limit
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          type: string
      - name: typologie
        in: query
        description: 'example: Décorations sur la voie publique'
        schema:
          type: string
      - name: categorie
        in: query
        description: 'example: Jardiniere'
        schema:
          type: string
      - name: code_postal
        in: query
        description: 'example: 75015'
        schema:
          type: string
      - name: annee_ouverture
        in: query
        description: 'example: 1978'
        schema:
          type: string
      - name: annee_renovation
        in: query
        description: 'example: 1995'
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/greenArea'
        500:
          description: Database error
          content: {}
  /espace-vert/{id}:
    get:
      tags:
      - paris
      summary: Get a green area by ID
      operationId: GreenAreaGET
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/greenArea'
        500:
          description: Database error
          content: {}
  /stations-ratp:
    get:
      tags:
      - ratp
      summary: Get all RATP stations
      operationId: StationsGET
      parameters:
      - name: limit
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          type: string
      - name: outside
        in: query
        schema:
          type: string
          enum:
          - "true"
          - "false"
      - name: type
        in: query
        schema:
          type: string
          enum:
          - train
          - RER
          - métro
          - bus
      - name: line
        in: query
        schema:
          type: string
          enum:
          - A
          - B
          - "1"
          - "2"
          - "3"
          - 3bis
          - "4"
          - "5"
          - "6"
          - "7"
          - 7bis
          - "8"
          - "9"
          - "10"
          - "11"
          - "12"
          - "13"
          - "14"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ratpStation'
        500:
          description: Database error
          content: {}
  /station-ratp/{id}:
    get:
      tags:
      - ratp
      summary: Get a RATP station by ID
      operationId: StationGET
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ratpStation'
        500:
          description: Database error
          content: {}
...
components:
  schemas:
...
    greenArea:
      required:
      - identifiant_espace_vert
      type: object
      properties:
        identifiant_espace_vert:
          type: integer
          format: int64
        feature:
          required:
          - geometry
          type: object
          properties:
            type:
              type: string
              example: Feature
            geometry:
              type: object
              properties:
                type:
                  type: string
                  example: MultiPolygon
                coordinates:
                  type: array
                  items:
                    type: array
                    items:
                      type: array
                      items:
                        type: array
                        items:
                            type: number
                            format: double
              required:
                - type
                - coordinates
            properties:
              type: object
              properties: {}
        nom_de_lespace_vert:
          type: string
          example: CENTRE SPORTIF DIDOT
        typologie_d_espace_vert:
          type: string
          example: Décorations sur la voie publique
        categorie:
          type: string
          example: Jardiniere
        adresse_numero:
          type: integer
        adresse_complement:
          type: string
          example: B
        adresse_type_voie:
          type: string
          example: RUE
        adresse_libelle_voie:
          type: string
          example: REIMS
        code_postal:
          type: integer
          format: int64
        surface_calculee:
          type: integer
          format: int64
        surface_totale_reelle:
          type: integer
          format: int64
        surface_horticole:
          type: integer
        presence_cloture:
          type: string
          example: Oui
        perimetre:
          type: number
          format: double
        annee_ouverture:
          type: integer
        annee_renovation:
          type: integer
        ancien_nom:
          type: string
          example: JARDIN DE LA DALLE DU PERIPHERIQUE PORTE DE VANVES
        annee_changement_nom:
          type: integer
        nombre_entites:
          type: integer
        ouverture_24h_sur24:
          type: string
          example: Non
        identifiant_division:
          type: integer
        idenifiant_atelier_horticole:
          type: integer
        ida3d_enb:
          type: string
          example: 194296 194300
        site_villes:
          type: string
          example: SV
        identifiant_equipement:
          type: string
          example: 4767 DJS
        competence:
          type: string
          example: CA
        url_plan:
          type: string
          example: http://p70-planet.apps.paris.mdp/Advitium/AppelPlan.asp?SystemId={1ED86C6E-0587-4045-A2D0-260669B88046}&Objectid=0
        geo_point:
          type: string
          example: ""
    ratpStation:
      type: object
      properties:
        identifiant:
            type: integer
            format: int64
        feature:
            type: object
            properties:
                type:
                    type: string
                    example: Feature
                geometry:
                    type: object
                    properties:
                        type:
                            type: string
                            example: Point
                        coordinates:
                            type: array
                            items:
                                type: array
                                items:
                                    type: number
                                    format: double
                        properties:
                            type: object
                    required:
                    - type
                    - coordinates
            required:
                - geometry
        name:
            type: string
            example: Châtelet
        address:
            type: string
            example: 25 rue des Volontaires - 75115
        coordinates:
            type: string
            example: 48.8414184427,2.30799364939
        train_outside:
            type: boolean
            default: false
        transport_type:
            type: string
            example: métro
        line:
            type: string
            example: 7bis
        line_index:
            type: integer
            format: int64    
...